+++ ffmpeg/src/main/java/org/bytedeco/ffmpeg/presets/avutil.java
@@ -63,7 +63,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
             "<libavutil/tx.h>", "<libavutil/version.h>", "<libavutil/macros.h>", "log_callback.h"},
             includepath = {"/usr/local/include/ffmpeg/", "/opt/local/include/ffmpeg/", "/usr/include/ffmpeg/"},
             link = "avutil@.59", compiler = {"default", "nodeprecated"}),
-        @Platform(value = "linux-x86", preload = {"va@.1", "drm@.2", "va-drm@.1"}, preloadpath = {"/usr/lib32/", "/usr/lib/"}),
+        @Platform(value = "linux-x86", preload = {"va@.2", "drm@.2", "va-drm@.2"}, preloadpath = {"/usr/lib32/", "/usr/lib/"}),
         @Platform(value = "linux-x86_64", preloadpath = {"/usr/lib64/", "/usr/lib/"}),
         @Platform(value = "macosx", preload = {"atomic@.1"},
                                     preloadpath = {"/usr/local/lib/gcc/13/", "/usr/local/lib/gcc/12/", "/usr/local/lib/gcc/11/", "/usr/local/lib/gcc/10/", "/usr/local/lib/gcc/9/",