+++ opencv/src/main/java/org/bytedeco/opencv/presets/opencv_cudafilters.java
@@ -36,12 +36,12 @@ import org.bytedeco.javacpp.tools.InfoMapper;
     value = {
         @Platform(
             include = "<opencv2/cudafilters.hpp>",
-            link = "opencv_cudafilters@.411",
+            link = "opencv_cudafilters@.412",
             extension = "-gpu"
         ),
         @Platform(
             value = "windows",
-            link = "opencv_cudafilters4110",
+            link = "opencv_cudafilters4120",
             extension = "-gpu"
         )
     },