+++ opencv/src/main/java/org/bytedeco/opencv/presets/opencv_cudaobjdetect.java
@@ -40,14 +40,14 @@ import org.bytedeco.javacpp.tools.InfoMapper;
     value = {
         @Platform(
             include = "<opencv2/cudaobjdetect.hpp>",
-            link = "opencv_cudaobjdetect@.411",
-            preload = "opencv_cudalegacy@.411",
+            link = "opencv_cudaobjdetect@.412",
+            preload = "opencv_cudalegacy@.412",
             extension = "-gpu"
         ),
         @Platform(
             value = "windows",
-            link = "opencv_cudaobjdetect4110",
-            preload = "opencv_cudalegacy4110",
+            link = "opencv_cudaobjdetect4120",
+            preload = "opencv_cudalegacy4120",
             extension = "-gpu"
         )
     },