+++ opencv/src/main/java/org/bytedeco/opencv/presets/opencv_cudastereo.java
@@ -33,12 +33,12 @@ import org.bytedeco.javacpp.tools.InfoMapper;
     value = {
         @Platform(
             include = "<opencv2/cudastereo.hpp>",
-            link = "opencv_cudastereo@.411",
+            link = "opencv_cudastereo@.412",
             extension = "-gpu"
         ),
         @Platform(
             value = "windows",
-            link = "opencv_cudastereo4110",
+            link = "opencv_cudastereo4120",
             extension = "-gpu"
         )
     },