+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 
  * Bundle libraries for VA-API 2.x with FFmpeg to avoid loading issues ([issue bytedeco/javacv#2340](https://github.com/bytedeco/javacv/issues/2340))
+ * Upgrade presets for OpenCV 4.12.0
 
 ### June 30, 2025 version 1.5.12
  * Add presets for the CUFILE module of CUDA